+++ src/deps.ts
@@ -31,3 +31,4 @@ export {
 	BaseHandler,
 	ConsoleHandler,
 } from 'https://deno.land/std@0.181.0/log/handlers.ts'
+export { crypto } from 'https://deno.land/std@0.186.0/crypto/mod.ts'